Convert all JS to TS and convert class components to functional components
